S3 파일 Persist 전략 끝
