+++ common/src/main/java/com/pluxity/file/strategy/storage/FilePersistenceContext.java
@@ -1,0 +1,11 @@
+package com.pluxity.file.strategy.storage;
+
+import lombok.Builder;
+
+@Builder
+public record FilePersistenceContext(
+        String filePath,
+        String newPath,
+        String contentType
+) {
+}
