+++ common/src/main/java/com/pluxity/file/strategy/storage/StorageStrategy.java
@@ -4,4 +4,5 @@ import org.springframework.web.multipart.MultipartFile;
 
 public interface StorageStrategy {
     FileProcessingContext save(MultipartFile multipartFile) throws Exception;
+    String persist(FilePersistenceContext context) throws Exception;
 }
